requirements.txt
aapt==1.0.2
adbutils==0.6.3
allure-pytest==2.8.6
allure-python-commons==2.8.6
androguard==3.4.0
apkutils2==1.0.0
Appium-Python-Client==0.48
asn1crypto==1.2.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
cigam==0.0.3
cli-helpers==1.2.1
Click==7.0
colorama==0.4.3
configobj==5.0.6
cryptography==2.8
cycler==0.10.0
ddt==1.2.1
decorator==4.4.1
Deprecated==1.2.7
deprecation==2.0.7
dicttoxml==1.7.4
enum34==1.1.6
et-xmlfile==1.0.1
hjson==3.0.1
humanize==0.5.1
idna==2.8
importlib-metadata==0.23
ipython==7.10.1
ipython-genutils==0.2.0
jdcal==1.4.1
Jinja2==2.10.3
kiwisolver==1.1.0
logzero==1.5.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==7.2.0
mycli==1.20.1
namedlist==1.7
networkx==2.4
numpy==1.17.2
openpyxl==3.0.1
packaging==19.2
pandas==0.25.3
pbr==5.4.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.0
progress==1.5
progressbar2==3.47.0
prompt-toolkit==3.0.2
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pycurl==7.43.0.5
pydot==1.4.1
pyelftools==0.26
Pygments==2.5.2
PyMySQL==0.9.3
pyparsing==2.4.5
pyperclip==1.7.0
PyQt5==5.13.2
PyQt5-sip==12.7.0
pytest==5.2.4
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-rerunfailures==7.0
pytest-variables==1.9.0
python-dateutil==2.8.1
python-magic-bin==0.4.14
python-utils==2.3.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-download==0.1.2
retry==0.9.2
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
star==0.1.2
stevedore==1.31.0
tabulate==0.8.6
terminaltables==3.1.0
traitlets==4.3.3
typesystem==0.2.4
uiautomator2==2.3.1
urllib3==1.25.6
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
whichcraft==0.6.1
wrapt==1.11.2
xmindparser==1.0.8
xmltodict==0.12.0
zipp==0.6.0
